Move includes out of extern "C" block
It is never recommended to include other files inside an extern "C"
block, as to not force any linkage on other header files. In this case,
an upcoming commit needs to include Wire.h from sensirion_i2c.h, which
would break on C++ constructs when inside the extern "C" block.